Business Consulting/requirements.txt
Flask==3.1.1
openai==1.82.0
pandas==2.2.3
numpy==2.2.6
python-dotenv==1.1.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Click==8.1.7
Jinja2==3.1.4
httpx==0.28.1
httpcore==1.0.9
idna==3.10
sniffio==1.3.1
tqdm==4.67.1
pydantic==2.11.5
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
prometheus-client==0.22.0
typing-extensions==4.13.2
typing-inspection==0.4.1
colorama==0.4.6
distro==1.9.0
jiter==0.10.0
MarkupSafe==2.1.5
six==1.17.0
